fixed issue with casting to jwtuserdetailsserver & put matching routes in different controller files
