+++ frontend/src/components/Search.js
@@ -3,17 +3,23 @@ import React,{ useState,useEffect }  from "react";
 import TutorialDataService from "../TutorialDataService";
 
 export default function Search(){
-    const [users,setUser] = useState({"key":"pppp"})
+    const [users,setUser] = useState({
+        "id": 0,
+        "firstName":"",
+        "lastName":"",
+        "email":"",
+        "summary":""
+    })
 // useEffect(()=>{async function getData(){console.log("get into0000");
 // TutorialDataService.getAllUsers()
 //     .then(res=> 
 //        { console.log("jjjjj");
 //  setUser(res.data); return res}).then(data=> (data))};
 // getData();},[])
-TutorialDataService.getAllUsers().then(
-    res=>setUser(res.data)
-)
-console.log(users)
+    TutorialDataService.getAllUsers().then(
+        res=>setUser(res.data)
+    )
+    console.log(users)
     return (<div>
        
         <div className="search-form"> <h3>Search Users</h3>
