+++ src/main/java/com/jlaurie/roommateFinder/RoommateFinderApplication.java
@@ -7,6 +7,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class RoommateFinderApplication {
 	
 	public static void main(String[] args) {
+		System.setProperty("spring.devtools.restart.enabled", "false");
 		SpringApplication.run(RoommateFinderApplication.class, args);
 	}
