+++ src/main/java/com/jlaurie/roommateFinder/jwt/JwtTokenUtil.java
@@ -89,7 +89,7 @@ public class JwtTokenUtil implements Serializable {
     }
 
     public Boolean validateToken(String token, UserDetails userDetails) {
-        JwtUserDetails user = (JwtUserDetails) userDetails;
+        UserDetails user =  userDetails;
         final String username = getUsernameFromToken(token);
         return (username.equals(user.getUsername()) && !isTokenExpired(token));
     }
