+++ src/main/resources/application.properties
@@ -10,6 +10,8 @@ jwt.refresh.token.uri=/refresh
 jwt.http.request.header=Authorization
 jwt.token.expiration.in.seconds=604800
 
+spring.devtools.restart.enabled = false
+
 # Hibernate Properties
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
